requirements.txt
requests==2.28.1
Flask==2.2.2
python-dotenv==0.21.0
Flask-HTTPAuth==4.7.0
Flask_Caching==2.0.1
pandas==1.5.1
